Sort out breadcrumb logic
